+++ message/name.go
@@ -16,6 +16,9 @@ type NameCollection interface {
 	// HasM returns true if NameOf(m) is in the container.
 	HasM(m dogma.Message) bool
 
+	// Len returns the number of names in the collection.
+	Len() int
+
 	// Range invokes fn once for each name in the container.
 	//
 	// Iteration stops when fn returns false or once fn has been invoked for all