+++ message/typeset.go
@@ -164,6 +164,11 @@ func (s TypeSet) IsEqual(o TypeSet) bool {
 	return true
 }
 
+// Len returns the number of types in the collection.
+func (s TypeSet) Len() int {
+	return len(s)
+}
+
 // Range invokes fn once for each type in the container.
 //
 // Iteration stops when fn returns false or once fn has been invoked for all